change int to long, add mappedsuperclass

+++ src/main/java/com/jesua/registration/dto/CourseDto.java
@@ -15,5 +15,5 @@ public class CourseDto {
     private int capacity;
     private UUID userId;
     private String image;
-    private int projectId;
+    private long projectId;
 }
