+++ src/main/java/com/jesua/registration/repository/CourseRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.time.Instant;
 import java.util.List;
 
-public interface CourseRepository extends JpaRepository<Course, Integer> {
+public interface CourseRepository extends JpaRepository<Course, Long> {
 
     List<Course> findByOpenTrue();
     List<Course> findByStartDateBetween(Instant start, Instant end);
